refactor(auth): group login/logout/whoami under `peppy auth` subcommand
