+++ crates/config-internal/src/peppy_config.rs
@@ -32,7 +32,7 @@ pub const PEPPY_CONFIG_FILE: &str = "peppy_config.json5";
 /// The backend resource-server URL for this build: the local dev backend in
 /// debug builds, the prod backend in release builds. The single source of truth
 /// for both the seeded `resource_servers` block and the built-in fallback the
-/// `peppy login` / `whoami` / `logout` commands resolve when no `--api-url` /
+/// `peppy auth login` / `whoami` / `logout` commands resolve when no `--api-url` /
 /// `PEPPY_API_URL` override is given.
 #[cfg(debug_assertions)]
 pub const DEFAULT_API_URL: &str = "http://127.0.0.1:3000";
@@ -182,7 +182,7 @@ const API_FIELD_SNIPPET: &str = const_format::concatcp!("    api: \"", DEFAULT_A
 /// CLI auth commands read this URL; the daemon ignores it but seeds and
 /// completes the block like every other knob.
 const RESOURCE_SERVERS_SECTION_SNIPPET: &str = const_format::concatcp!(
-    r#"  // Backend resource-server URL the `peppy login` / `whoami` / `logout`
+    r#"  // Backend resource-server URL the `peppy auth login` / `whoami` / `logout`
   // commands talk to. Baked in at compile time (the dev backend in debug
   // builds, prod in release); --api-url / PEPPY_API_URL override it at runtime.
   resource_servers: {
